# artifacts/results/openapi.json
{
  "components": {
    "schemas": {
      "AffectPoint": {
        "properties": {
          "arousal": {
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Arousal",
            "type": "number"
          },
          "valence": {
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Valence",
            "type": "number"
          }
        },
        "required": [
          "valence",
          "arousal"
        ],
        "title": "AffectPoint",
        "type": "object"
      },
      "DayNight": {
        "enum": [
          "Day",
          "Night"
        ],
        "title": "DayNight",
        "type": "string"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "model_status": {
            "title": "Model Status",
            "type": "string"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "const": "alive",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "status",
          "ready",
          "model_status"
        ],
        "title": "HealthResponse",
        "type": "object"
      },
      "MetaResponse": {
        "properties": {
          "artifact_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Version"
          },
          "data_manifest_sha256": {
            "title": "Data Manifest Sha256",
            "type": "string"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "model_status": {
            "title": "Model Status",
            "type": "string"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "data_manifest_sha256",
          "artifact_version",
          "model_status",
          "ready",
          "limitations"
        ],
        "title": "MetaResponse",
        "type": "object"
      },
      "OptimizeCandidate": {
        "properties": {
          "prediction": {
            "$ref": "#/components/schemas/PredictionPayload"
          },
          "room": {
            "$ref": "#/components/schemas/RoomInput"
          },
          "support": {
            "$ref": "#/components/schemas/SupportPayload"
          }
        },
        "required": [
          "room",
          "prediction",
          "support"
        ],
        "title": "OptimizeCandidate",
        "type": "object"
      },
      "OptimizeRequest": {
        "properties": {
          "budget": {
            "default": 500,
            "maximum": 5000.0,
            "minimum": 1.0,
            "title": "Budget",
            "type": "integer"
          },
          "day_or_night": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DayNight"
              },
              {
                "type": "null"
              }
            ]
          },
          "n_candidates": {
            "default": 5,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "N Candidates",
            "type": "integer"
          },
          "seed": {
            "default": 2718,
            "title": "Seed",
            "type": "integer"
          },
          "space_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpaceType"
              },
              {
                "type": "null"
              }
            ]
          },
          "target": {
            "$ref": "#/components/schemas/AffectPoint"
          }
        },
        "required": [
          "target"
        ],
        "title": "OptimizeRequest",
        "type": "object"
      },
      "OptimizeResponse": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/OptimizeCandidate"
            },
            "title": "Candidates",
            "type": "array"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "model_status": {
            "title": "Model Status",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "samples_evaluated": {
            "title": "Samples Evaluated",
            "type": "integer"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "ok",
              "empty"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "status",
          "model_status",
          "candidates",
          "samples_evaluated",
          "reason",
          "limitations"
        ],
        "title": "OptimizeResponse",
        "type": "object"
      },
      "PredictRequest": {
        "properties": {
          "room": {
            "$ref": "#/components/schemas/RoomInput"
          },
          "target": {
            "$ref": "#/components/schemas/AffectPoint"
          }
        },
        "required": [
          "room",
          "target"
        ],
        "title": "PredictRequest",
        "type": "object"
      },
      "PredictResponse": {
        "properties": {
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "model_status": {
            "title": "Model Status",
            "type": "string"
          },
          "prediction": {
            "$ref": "#/components/schemas/PredictionPayload"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "const": "ok",
            "title": "Status",
            "type": "string"
          },
          "support": {
            "$ref": "#/components/schemas/SupportPayload"
          }
        },
        "required": [
          "schema_version",
          "status",
          "model_status",
          "prediction",
          "support",
          "limitations"
        ],
        "title": "PredictResponse",
        "type": "object"
      },
      "PredictionPayload": {
        "properties": {
          "arousal": {
            "title": "Arousal",
            "type": "number"
          },
          "neuro_score": {
            "title": "Neuro Score",
            "type": "number"
          },
          "projected": {
            "title": "Projected",
            "type": "boolean"
          },
          "raw_arousal": {
            "title": "Raw Arousal",
            "type": "number"
          },
          "raw_valence": {
            "title": "Raw Valence",
            "type": "number"
          },
          "valence": {
            "title": "Valence",
            "type": "number"
          }
        },
        "required": [
          "valence",
          "arousal",
          "raw_valence",
          "raw_arousal",
          "projected",
          "neuro_score"
        ],
        "title": "PredictionPayload",
        "type": "object"
      },
      "RoomInput": {
        "additionalProperties": false,
        "description": "Independent variables only; missing optional observations are not zeros.",
        "properties": {
          "cct": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cct"
          },
          "day_or_night": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DayNight"
              },
              {
                "type": "null"
              }
            ]
          },
          "daylight_factor": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Daylight Factor"
          },
          "door_area": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Door Area"
          },
          "height": {
            "exclusiveMinimum": 0.0,
            "maximum": 30.0,
            "title": "Height",
            "type": "number"
          },
          "illuminance": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Illuminance"
          },
          "length": {
            "exclusiveMinimum": 0.0,
            "maximum": 100.0,
            "title": "Length",
            "type": "number"
          },
          "num_doors": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Doors"
          },
          "num_windows": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Windows"
          },
          "space_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SpaceType"
              },
              {
                "type": "null"
              }
            ]
          },
          "walkable_floor_area": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Walkable Floor Area"
          },
          "width": {
            "exclusiveMinimum": 0.0,
            "maximum": 100.0,
            "title": "Width",
            "type": "number"
          },
          "window_area": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Area"
          }
        },
        "required": [
          "length",
          "width",
          "height"
        ],
        "title": "RoomInput",
        "type": "object"
      },
      "SpaceType": {
        "enum": [
          "General",
          "Bedroom",
          "Living Room",
          "Workplace",
          "Classroom",
          "Cafeteria"
        ],
        "title": "SpaceType",
        "type": "string"
      },
      "SupportPayload": {
        "properties": {
          "nearest_room_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nearest Room Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "standardized_distance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Standardized Distance"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Threshold"
          }
        },
        "required": [
          "status",
          "nearest_room_id",
          "standardized_distance",
          "threshold"
        ],
        "title": "SupportPayload",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Predictive Atmospheres Research API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/health": {
      "get": {
        "operationId": "health_v1_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/v1/meta": {
      "get": {
        "operationId": "metadata_v1_meta_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetaResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Metadata"
      }
    },
    "/v1/optimize": {
      "post": {
        "operationId": "optimize_v1_optimize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OptimizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptimizeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Optimize"
      }
    },
    "/v1/predict": {
      "post": {
        "operationId": "predict_v1_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Predict"
      }
    }
  }
}